requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow>=10.4.0
django-filter==23.5
python-decouple==3.8
django-filter==23.5 